refactor: review fixes for email validation, password reset, settings split
- Extract shared sendEmail helper in email client (Resend wrapper)
- Add validation utils (isValidEmail, MIN_PASSWORD_LENGTH) and persistLocale helper
- Rename Better Auth client sendVerificationEmail export to requestVerificationEmail (aliased destructure) and centralize resend flow in useResendVerification hook
- Use hook + utils in login/register/forgot/reset pages; register now persists en locale unconditionally, matching login
- Split AccountSettings into NameForm + PasswordChangeForm composition
- Drop redundant autoSignInAfterVerification (Better Auth default true)
