+++ delivery/webapp/src/lib/auth-client.ts
@@ -13,5 +13,5 @@ export const {
   resetPassword,
   changePassword,
   updateUser,
-  sendVerificationEmail,
+  sendVerificationEmail: requestVerificationEmail,
 } = authClient;
