+++ delivery/webapp/src/lib/persist-locale.ts
@@ -1,0 +1,11 @@
+export async function persistLocale(locale: string): Promise<void> {
+  try {
+    await fetch("/api/settings", {
+      method: "PUT",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify({ locale }),
+    });
+  } catch {
+    // best-effort
+  }
+}
