+++ delivery/webapp/src/lib/validation.ts
@@ -1,0 +1,5 @@
+export const MIN_PASSWORD_LENGTH = 8;
+
+export function isValidEmail(email: string): boolean {
+  return /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email);
+}
