+++ delivery/webapp/src/test/app/settings-save-error.test.tsx
@@ -21,7 +21,7 @@ vi.mock("@/lib/auth-client", () => ({
   signOut: () => Promise.resolve(),
   updateUser: vi.fn(() => Promise.resolve({ data: null, error: null })),
   changePassword: vi.fn(() => Promise.resolve({ data: null, error: null })),
-  sendVerificationEmail: vi.fn(() => Promise.resolve({ data: null, error: null })),
+  requestVerificationEmail: vi.fn(() => Promise.resolve({ data: null, error: null })),
   useSession: () => ({
     data: { user: { id: "1", name: "Test User", email: "user@example.com" } },
     isPending: false,
