+++ delivery/webapp/src/test/app/settings-signout.test.tsx
@@ -23,7 +23,7 @@ vi.mock("@/lib/auth-client", () => ({
   signOut: (...args: unknown[]) => mockSignOut(...args),
   updateUser: vi.fn(() => Promise.resolve({ data: null, error: null })),
   changePassword: vi.fn(() => Promise.resolve({ data: null, error: null })),
-  sendVerificationEmail: vi.fn(() => Promise.resolve({ data: null, error: null })),
+  requestVerificationEmail: vi.fn(() => Promise.resolve({ data: null, error: null })),
   useSession: () => ({
     data: { user: { id: "1", name: "Test User", email: "user@example.com" } },
     isPending: false,
